Scopable module for api should be shared
