+++ app/controllers/api/customer/cart/checkouts_controller.rb
@@ -2,6 +2,7 @@ module Api
   module Customer
     module Cart
       class CheckoutsController < Api::Customer::BaseController
+        include Api::BaseController::Scopable
         include Api::BaseController::Resourceable
 
         resource_description do
