+++ app/controllers/api/customer/wishlists_controller.rb
@@ -1,6 +1,7 @@
 module Api
   module Customer
     class WishlistsController < BaseController
+      include Api::BaseController::Scopable
       include Api::BaseController::Resourceable
       include Api::Customer::BaseController::ResourceCollectionable
